+++ paifeaturestore-20230621/alibabacloud_paifeaturestore20230621/models/_list_llmconfigs_response_body.py
@@ -83,11 +83,13 @@ class ListLLMConfigsResponseBodyLLMConfigs(DaraModel):
         base_url: str = None,
         batch_size: int = None,
         embedding_dimension: int = None,
+        enable_fusion: bool = None,
         gmt_create_time: str = None,
         gmt_modified_time: str = None,
         llmconfig_id: str = None,
         max_tokens: int = None,
         model: str = None,
+        model_type: str = None,
         name: str = None,
         resource_group_id: str = None,
         rps: int = None,
@@ -97,11 +99,13 @@ class ListLLMConfigsResponseBodyLLMConfigs(DaraModel):
         self.base_url = base_url
         self.batch_size = batch_size
         self.embedding_dimension = embedding_dimension
+        self.enable_fusion = enable_fusion
         self.gmt_create_time = gmt_create_time
         self.gmt_modified_time = gmt_modified_time
         self.llmconfig_id = llmconfig_id
         self.max_tokens = max_tokens
         self.model = model
+        self.model_type = model_type
         self.name = name
         self.resource_group_id = resource_group_id
         self.rps = rps
@@ -127,6 +131,9 @@ class ListLLMConfigsResponseBodyLLMConfigs(DaraModel):
         if self.embedding_dimension is not None:
             result['EmbeddingDimension'] = self.embedding_dimension
 
+        if self.enable_fusion is not None:
+            result['EnableFusion'] = self.enable_fusion
+
         if self.gmt_create_time is not None:
             result['GmtCreateTime'] = self.gmt_create_time
 
@@ -141,6 +148,9 @@ class ListLLMConfigsResponseBodyLLMConfigs(DaraModel):
 
         if self.model is not None:
             result['Model'] = self.model
+
+        if self.model_type is not None:
+            result['ModelType'] = self.model_type
 
         if self.name is not None:
             result['Name'] = self.name
@@ -170,6 +180,9 @@ class ListLLMConfigsResponseBodyLLMConfigs(DaraModel):
         if m.get('EmbeddingDimension') is not None:
             self.embedding_dimension = m.get('EmbeddingDimension')
 
+        if m.get('EnableFusion') is not None:
+            self.enable_fusion = m.get('EnableFusion')
+
         if m.get('GmtCreateTime') is not None:
             self.gmt_create_time = m.get('GmtCreateTime')
 
@@ -184,6 +197,9 @@ class ListLLMConfigsResponseBodyLLMConfigs(DaraModel):
 
         if m.get('Model') is not None:
             self.model = m.get('Model')
+
+        if m.get('ModelType') is not None:
+            self.model_type = m.get('ModelType')
 
         if m.get('Name') is not None:
             self.name = m.get('Name')