+++ paifeaturestore-20230621/alibabacloud_paifeaturestore20230621/models/_update_llmconfig_request.py
@@ -11,8 +11,10 @@ class UpdateLLMConfigRequest(DaraModel):
         base_url: str = None,
         batch_size: int = None,
         embedding_dimension: int = None,
+        enable_fusion: bool = None,
         max_tokens: int = None,
         model: str = None,
+        model_type: str = None,
         name: str = None,
         rps: int = None,
     ):
@@ -22,10 +24,12 @@ class UpdateLLMConfigRequest(DaraModel):
         self.base_url = base_url
         self.batch_size = batch_size
         self.embedding_dimension = embedding_dimension
+        self.enable_fusion = enable_fusion
         # This parameter is required.
         self.max_tokens = max_tokens
         # This parameter is required.
         self.model = model
+        self.model_type = model_type
         # This parameter is required.
         self.name = name
         # This parameter is required.
@@ -51,11 +55,17 @@ class UpdateLLMConfigRequest(DaraModel):
         if self.embedding_dimension is not None:
             result['EmbeddingDimension'] = self.embedding_dimension
 
+        if self.enable_fusion is not None:
+            result['EnableFusion'] = self.enable_fusion
+
         if self.max_tokens is not None:
             result['MaxTokens'] = self.max_tokens
 
         if self.model is not None:
             result['Model'] = self.model
+
+        if self.model_type is not None:
+            result['ModelType'] = self.model_type
 
         if self.name is not None:
             result['Name'] = self.name
@@ -79,11 +89,17 @@ class UpdateLLMConfigRequest(DaraModel):
         if m.get('EmbeddingDimension') is not None:
             self.embedding_dimension = m.get('EmbeddingDimension')
 
+        if m.get('EnableFusion') is not None:
+            self.enable_fusion = m.get('EnableFusion')
+
         if m.get('MaxTokens') is not None:
             self.max_tokens = m.get('MaxTokens')
 
         if m.get('Model') is not None:
             self.model = m.get('Model')
+
+        if m.get('ModelType') is not None:
+            self.model_type = m.get('ModelType')
 
         if m.get('Name') is not None:
             self.name = m.get('Name')